+++ include/databento/live_threaded.hpp
@@ -65,8 +65,8 @@ class LiveThreaded {
                  SType stype_in, UnixNanos start);
   void Subscribe(const std::vector<std::string>& symbols, Schema schema,
                  SType stype_in, const std::string& start);
-  void Subscribe(const std::vector<std::string>& symbols, Schema schema,
-                 SType stype_in, bool use_snapshot);
+  void SubscribeWithSnapshot(const std::vector<std::string>& symbols,
+                             Schema schema, SType stype_in);
   // Notifies the gateway to start sending messages for all subscriptions.
   // `metadata_callback` will be called exactly once, before any calls to
   // `record_callback`. `record_callback` will be called for records from all